Improve Dockerfile to use local JAR and include an ENTRYPOINT.
The use of an ENTRYPOINT allows using the container like an application:

```
docker run searlelab/encyclopedia:$VERSION -Xmx4G <arguments>
```

Switching to use the local JAR during `docker build` simplifies the
container setup and enables developers to build the container with
local changes. Note that this requires you to run `mvn package -PbuildEncyclopedia`
prior to `docker build`, and you must specify the correct value
for `--build-arg VERSION=...`.

When building on Bitbucket Pipelines the JAR will be available in the
local context due to the `artifacts` defined in the `Build and Test`
step.

Note that the specification of ENTRYPOINT is complicated by the need to
expand the $VERSION variable while also supporting the specification of
parameters via CMD or `docker run`. See comments in `Dockerfile` about
the approach taken.