Prevent buggy initial round of Percolator training
This bug has been latent in the code since May 2020, when I added
support for the Percolator `--trainFDR` option. This parameter's
default value is `0.0`, indicating that the same threshold as the
`--testFDR` option should be used. However, it appears that the
`--train-fdr-initial` option (which we don't specify) defaults to
the `--trainFDR` value directly, leading to bad initial feature
selection. This behavior does not affect users who specify the
`-percolatorTrainingThreshold` parameter, or use a pre-3.05 version
of Percolator.

The primary indication of the bug is that Percolator's log output
will show that it was unable to determine an initial search feature
because no peptides could be found at "q<0".

Another indication (though other factors interfere with this) is that
integration tests will sometimes fail with zero peptides identified.